+++ design-system/form/src/__tests__/unit/server-side-hydrate.tsx
@@ -12,7 +12,7 @@ afterEach(() => {
 });
 
 test('should ssr then hydrate form correctly', async () => {
-  const [example] = await getExamplesFor('form');
+  const [example] = await getExamplesFor('@atlaskit/form');
   const Example = require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
 
   const elem = document.createElement('div');