+++ media/media-core/CHANGELOG.md
@@ -921,6 +921,7 @@ const config: MediaClientConfig = {
 * breaking; Api for DataUriService has changed
 * before:
 * ```typescript
+
   ```
 
 * interface fetchImageDataUri(
@@ -935,6 +936,7 @@ const config: MediaClientConfig = {
 
 * after:
 * ```typescript
+
   ```
 
 * interface FetchImageOptions {